Switching context now correctly initialize api client
Before this patch, switching context within the repl would not reinitialize the api
client with the unintended effect for commands to keep targeting the
previous context.

I ventured with some refactoring but didn't push too far since it would
have required too much time.
